Error rendering Product in Index

+++ entities/product.go
@@ -6,7 +6,7 @@ type Product struct {
 	Id          uint
 	Name        string
 	Category    Category
-	Stock       uint
+	Stock       int64
 	Description string
 	CreatedAt   time.Time
 	UpdatedAt   time.Time
